refactor(Badge): split component files
